Move MqttRoot to its own source file
