+++ main/peripherals/environment/Ds18B20SoilSensor.hpp
@@ -10,7 +10,6 @@
 #include <peripherals/Peripheral.hpp>
 #include <kernel/Configuration.hpp>
 #include <kernel/Component.hpp>
-#include <kernel/mqtt/MqttDriver.hpp>
 #include <peripherals/SinglePinDeviceConfig.hpp>
 
 using namespace farmhub::kernel;
